pyproject.toml
[project]
name = "preseal"
version = "0.1.0"
description = "Pre-deployment security testing for AI agents. Find structural vulnerabilities before production."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3",
    "typer>=0.12",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
langgraph = ["langgraph>=0.2"]
dev = ["pytest>=8.0", "ruff>=0.4"]

[project.scripts]
preseal = "preseal.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
